Restructure cache interfaces and cleanup cache implementation
