+++ rosbag2_cpp/include/rosbag2_cpp/cache/cache_consumer.hpp
@@ -23,7 +23,7 @@
 #include <vector>
 
 #include "rosbag2_cpp/cache/message_cache.hpp"
-#include "rosbag2_cpp/cache_interfaces/base_message_cache_interface.hpp"
+#include "rosbag2_cpp/cache/message_cache_interface.hpp"
 
 // This is necessary because of using stl types here. It is completely safe, because
 // a) the member is not accessible from the outside
@@ -62,10 +62,10 @@ class ROSBAG2_CPP_PUBLIC CacheConsumer
 {
 public:
   using consume_callback_function_t = std::function<void (const
-      std::vector<rosbag2_cpp::buffer_element_t> &)>;
+      std::vector<buffer_element_t> &)>;
 
   CacheConsumer(
-    std::shared_ptr<rosbag2_cpp::cache_interfaces::BaseMessageCacheInterface> message_cache,
+    std::shared_ptr<MessageCacheInterface> message_cache,
     consume_callback_function_t consume_callback);
 
   ~CacheConsumer();
@@ -77,7 +77,7 @@ public:
   void change_consume_callback(consume_callback_function_t callback);
 
 private:
-  std::shared_ptr<rosbag2_cpp::cache_interfaces::BaseMessageCacheInterface> message_cache_;
+  std::shared_ptr<MessageCacheInterface> message_cache_;
   consume_callback_function_t consume_callback_;
 
   /// Write buffer data to a storage
