+++ rosbag2_cpp/include/rosbag2_cpp/cache/circular_message_cache.hpp
@@ -20,8 +20,8 @@
 #include <string>
 
 #include "rosbag2_cpp/cache/message_cache_circular_buffer.hpp"
-#include "rosbag2_cpp/cache_interfaces/base_message_cache_interface.hpp"
-#include "rosbag2_cpp/cache_interfaces/base_cache_buffer_interface.hpp"
+#include "rosbag2_cpp/cache/message_cache_interface.hpp"
+#include "rosbag2_cpp/cache/cache_buffer_interface.hpp"
 #include "rosbag2_cpp/visibility_control.hpp"
 
 #include "rosbag2_storage/serialized_bag_message.hpp"
@@ -40,7 +40,7 @@ namespace cache
 {
 
 class ROSBAG2_CPP_PUBLIC CircularMessageCache
-  : public rosbag2_cpp::cache_interfaces::BaseMessageCacheInterface
+  : public MessageCacheInterface
 {
 public:
   explicit CircularMessageCache(uint64_t max_buffer_size);
@@ -49,8 +49,7 @@ public:
   void push(std::shared_ptr<const rosbag2_storage::SerializedBagMessage> msg) override;
 
   /// get current buffer to consume
-  std::shared_ptr<rosbag2_cpp::cache_interfaces::BaseCacheBufferInterface>
-  consumer_buffer() override;
+  std::shared_ptr<CacheBufferInterface> consumer_buffer() override;
 
   /// Swap the primary and secondary buffer before consumption.
   /// NOTE: consumer_buffer() should be called sequentially after
