+++ rosbag2_cpp/include/rosbag2_cpp/cache/message_cache_interface.hpp
@@ -12,29 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef ROSBAG2_CPP__CACHE_INTERFACES__BASE_MESSAGE_CACHE_INTERFACE_HPP_
-#define ROSBAG2_CPP__CACHE_INTERFACES__BASE_MESSAGE_CACHE_INTERFACE_HPP_
+#ifndef ROSBAG2_CPP__CACHE__MESSAGE_CACHE_INTERFACE_HPP_
+#define ROSBAG2_CPP__CACHE__MESSAGE_CACHE_INTERFACE_HPP_
 
 #include <memory>
 
 #include "rosbag2_cpp/visibility_control.hpp"
-#include "rosbag2_cpp/cache_interfaces/base_cache_buffer_interface.hpp"
+#include "rosbag2_cpp/cache/cache_buffer_interface.hpp"
 #include "rosbag2_storage/serialized_bag_message.hpp"
 
 namespace rosbag2_cpp
 {
-namespace cache_interfaces
+namespace cache
 {
 
-class ROSBAG2_CPP_PUBLIC BaseMessageCacheInterface
+class ROSBAG2_CPP_PUBLIC MessageCacheInterface
 {
 public:
-  virtual ~BaseMessageCacheInterface() {}
+  virtual ~MessageCacheInterface() {}
 
   virtual void push(std::shared_ptr<const rosbag2_storage::SerializedBagMessage> msg) = 0;
 
-  virtual std::shared_ptr<rosbag2_cpp::cache_interfaces::BaseCacheBufferInterface>
-  consumer_buffer() = 0;
+  virtual std::shared_ptr<CacheBufferInterface> consumer_buffer() = 0;
 
   virtual void swap_buffers() = 0;
 
@@ -45,7 +44,7 @@ public:
   virtual void finalize() {}
 };
 
-}  // namespace cache_interfaces
+}  // namespace cache
 }  // namespace rosbag2_cpp
 
-#endif  // ROSBAG2_CPP__CACHE_INTERFACES__BASE_MESSAGE_CACHE_INTERFACE_HPP_
+#endif  // ROSBAG2_CPP__CACHE__MESSAGE_CACHE_INTERFACE_HPP_
