+++ rosbag2_cpp/src/rosbag2_cpp/cache/cache_consumer.cpp
@@ -23,7 +23,7 @@ namespace cache
 {
 
 CacheConsumer::CacheConsumer(
-  std::shared_ptr<rosbag2_cpp::cache_interfaces::BaseMessageCacheInterface> message_cache,
+  std::shared_ptr<MessageCacheInterface> message_cache,
   consume_callback_function_t consume_callback)
 : message_cache_(message_cache),
   consume_callback_(consume_callback)
